updated files to work with both linux and WSL

+++ stop.sh
@@ -16,10 +16,23 @@ echo -n "Stopping IoT Device Simulation..."
 sudo pkill -f LightSimulation.py
 echo "done."
 
-echo -n "Stopping Mosquitto..."
-sudo systemctl stop mosquitto
-echo "done."
+if command -v systemctl &> /dev/null
+then
+    echo "Running on native Linux"
+    echo -n "Stopping Mosquitto..."
+    sudo systemctl stop mosquitto
+    echo "done."
 
-echo -n "Stopping MongoDB..."
-sudo sudo systemctl stop mongod
-echo "done."
+    echo -n "Stopping MongoDB..."
+    sudo systemctl stop mongod
+    echo "done."
+else
+    echo "Running on Windows Subsystem for Linux"
+    echo -n "Stopping Mosquitto..."
+    sudo service mosquitto stop
+    echo "done."
+
+    echo -n "Stopping MongoDB..."
+    sudo service mongod stop
+    echo "done."
+fi
